requirements.txt
absl-py==2.1.0
cloudpickle==3.1.1
colorama==0.4.6
etils==1.12.0
Farama-Notifications==0.0.4
filelock==3.13.1
fsspec==2024.6.1
glfw==2.8.0
grpcio==1.70.0
gymnasium==1.1.0
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
mujoco==3.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.0.0
protobuf==5.29.3
PyOpenGL==3.1.9
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
matplotlib
psutil
wandb
hydra-core
omegaconf
transformers
diffusers
accelerate
safetensors
